requirements.txt
python-pptx>=1.0.2
boto3>=1.40.1
botocore>=1.40.1
python-docx>=1.1.2
beautifulsoup4>=4.13.4
openpyxl>=3.1.2
fastmcp>=2.5.2
pydantic>=2.11.5
PyYAML
pystache>=0.6.5
google-cloud-storage>=2.18.2
azure-storage-blob>=12.23.1
